@@ -11,27 +11,28 @@
     "type": "git",
     "url": "https://github.com/GLeeWei/cc-sdk.git"
   },
+  "type": "module",
   "sdkName": "ccx",
   "author": "Lee",
   "license": "ISC",
   "devDependencies": {
-    "@babel/core": "^7.18.6",
-    "@babel/preset-env": "^7.18.6",
-    "@babel/preset-typescript": "^7.18.6",
-    "@rollup/plugin-babel": "^5.3.1",
-    "@rollup/plugin-commonjs": "^22.0.1",
-    "@rollup/plugin-node-resolve": "^13.3.0",
-    "del": "^6.1.1",
-    "fs-extra": "^10.1.0",
+    "@babel/core": "^7.22.15",
+    "@babel/preset-env": "^7.22.15",
+    "@babel/preset-typescript": "^7.22.15",
+    "@rollup/plugin-babel": "^6.0.3",
+    "@rollup/plugin-commonjs": "^25.0.4",
+    "@rollup/plugin-node-resolve": "^15.2.1",
+    "@rollup/plugin-terser": "^0.4.3",
+    "del": "^7.1.0",
+    "fs-extra": "^11.1.1",
+    "gen-dts": "git+https://github.com/meetleev/gen-dts.git",
     "gulp": "^4.0.2",
     "gulp-babel": "^8.0.0",
     "gulp-concat": "^2.6.1",
-    "gulp-replace": "^1.1.3",
+    "gulp-replace": "^1.1.4",
     "gulp-wrap": "^0.15.0",
-    "rollup": "^2.76.0",
-    "rollup-plugin-terser": "^7.0.2",
-    "tslib": "^2.4.0",
-    "typescript": "^4.7.4",
-    "gen-dts": "git+https://github.com/meetleev/gen-dts.git"
+    "rollup": "^3.28.1",
+    "tslib": "^2.6.2",
+    "typescript": "^5.2.2"
   }
 }
